Add size, margins and rects queries which write out components separately
These are appended with _xy, _ltrb and _xywh respectively. Like the
functions which set values by passing the components separately instead of
in a lay_vec4 or a lay_vec2, these functions will write out the components
to the addresses passed as parameters instead of returning a lay_vec4 or a
lay_vec2 by value.

These are just helper wrappers in case your code doesn't want to deal with
the lay_vecX types.