Change to use __restrict instead of restrict when compiling as C++ in GCC
`restrict` is correctly not recognized as a valid identifier in C++ mode.
GCC provides `__restrict` which can be used in this case. We still prefer
to use `restrict` when compiling as C99. Also, we add a definition for
`LAY_RESTRICT` as `restrict` even if we don't recognize what compiler is
being used, as long as it's not C++, since `restrict` is part of C99.

Fixes #5