Fix left margin unaccounted in lay_arrange_overlay
lay_arrange_overlay() did not account for the left (or top) child margin
in the LAY_RIGHT fill case for a child. This could cause it to have an
incorrect offset if the child actually had a margin and LAY_RIGHT (or
bottom) was specified.

This commit also includes two test cases that will fail before the
patch, and succeed after the patch.

The user who reported the issue says that there may remain an issue in
lay_arrange_overlay_squeezed_range, but it has not yet been tested and
this commit does not address any possible issue there.